// src/types/statistics.ts
export interface ChartData {
  labels: string[];
  data: number[];
}

export interface RankSchoolType {
  schoolName: string;
  userCnt: number;
}

export interface RankLanguageType {
  languageAbbr: string;
  userCnt: number;
}

export interface RankHopeLanguageType {
  hopeLanguageAbbr: string;
  userCnt: number;
}

export interface RankVisitedLanguageType {
  languageAbbr: string;
  userCnt: number;
}